backend/requirements.txt
Django==2.0.6
djangorestframework==3.9.2
djangorestframework-jwt==1.11.0
PyMySQL==0.8.1
drf-yasg==1.15.0
swagger-spec-validator==2.1.0
flex==6.13.2
httpie==0.9.9
PyJWT==1.6.4
gunicorn==19.8.1
pyotp==2.2.6
django-redis==4.10.0
django-filter==1.1.0
pylint==1.9.2
autopep8==1.3.5
msgpack==0.5.6
django-cors-headers==2.3.0
django-crontab==0.7.1
django-rest-swagger==2.2.0
django-softdelete-it==0.1
django-aliyun-oss2-storage==0.1.5
xlwt==1.3.0
redis==3.2.1
drf-renderer-xlsx==0.3.2
vine==1.1.3
django-celery-beat==1.5.0
celery==4.3.0
# drf-extensions==0.5.1 注意 可能会应为drf版本引起不兼容，使用drf3.9没问题
# pip install -U "celery[redis]"  # 使用redis作为Broker
# celery -A base_django_api worker -l info # 启动celery